+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Version log for Spekkio CLI.
 
+## 0.1.0
+### New
+- Added shell script to build app
+### Modified
+- Changed config to parse / initialize a spekkiofile instead of .env
+### Removed
+- Removed `.env.example`
+
 ## 0.0.0
 ### New
 - Initial app CLI!